pyproject.toml
# SPDX-FileCopyrightText: 2023 Kevin Meagher
#
# SPDX-License-Identifier: GPL-3.0-or-later

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[project]
authors = [{name = "Kevin Meagher"}]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Utilities"
]
dependencies = [
  "backports.zoneinfo; python_version < '3.9'",
  "babel",
  "tabulate",
  "tzlocal",
  "xdg-base-dirs",
  "termcolor",
  "tomli; python_version < '3.11'"
]
dynamic = ["version", "description"]
license = {file = "COPYING"}
name = "multizone"

[project.optional-dependencies]
dev = ["pre-commit"]
mypy = ["types-tabulate", "types-tzlocal", "types-termcolor"]
test = ["pytest", "pytest-cov"]

[project.scripts]
multizone = "multizone:main"

[tool.black]
line-length = 108
target-version = ['py38']

[tool.coverage.run]
include = ["multizone.py"]

[tool.pylint.format]
max-line-length = "108"

[tool.pylint.message_control]
disable = "R0912"

[tool.ruff]
fixable = ["D", "I", "COM"]
ignore = [
  "S101",  # assert
  "D213",  # multi-line-summary-second-line incompatible with multi-line-summary-first-line
  "D203"  # one-blank-line-before-class" incompatible with no-blank-line-before-class
]
line-length = 108
select = ["ALL"]
target-version = "py38"

[tool.ruff.per-file-ignores]
"test_mz.py" = ["PLR2004"]
